replace secondary logs with one single costructure
Merge "next index", "reference", "author" and "index shift" into one BTreeMap, as an optimization suggested in the research paper. The integer key of the BTreeMap consists of two parts. Its 2 most significant bits denote the type of the metadata stored, and the remaining 30 bits are for the actual key stored. struct `Costructures` encapsulates the BTreeMap and provides getters and setters that reimplement the logic of `RangeFromMap` and `OffsetMap`.

For data compaction, all metadata are stored as usize instead of optionals or enums. That also means `Author` can no longer be an arbitrary type but a usize "convertible". As a result the random test is commented out for its incompatible `&str` keys.